Updated the products folder to include components such as GetProductReviews, GetProductByPrice, and GetProductByTags, and added a utility component.
